requirements.txt
torch
torchvision
numpy>=1.19.2
pandas>=1.1.3
matplotlib>=3.3.2
opencv-python>=4.4.0.46
Pillow>=8.0.1
tqdm>=4.51.0
scikit-learn>=0.23.2
tensorboard>=2.4.0
PyYAML
